+++ src/background/Background.ts
@@ -4,9 +4,7 @@ import { type PaymentFlowService } from '@/background/paymentFlow'
 import { setStorageDefaultData } from '@/utils/storage'
 import { EXTMessage } from '@/utils/types'
 
-import { generateKeysHandler } from './installHandlers'
-import { BrowserEventsService, EventsService } from './services'
-import { tabChangeHandler, tabUpdateHandler } from './tabHandlers'
+import { BrowserEventsService, EventsService, OpenPaymentsService } from './services'
 
 class Background {
   // TO DO: remove these from background into storage or state & use injection
@@ -18,6 +16,7 @@ class Background {
     private browser: Browser,
     private eventsService: EventsService,
     private browserEventsService: BrowserEventsService,
+    private openPaymentsService: OpenPaymentsService,
   ) {
     setStorageDefaultData()
   }
@@ -27,6 +26,11 @@ class Background {
       switch (message.type) {
         case 'GET_STORAGE_DATA':
           return await this.eventsService.getStorageData()
+
+        // case 'SUBMIT_FORM':
+        //   await this.openPaymentsService.initClient('https://ilp.rafiki.money/radu')
+        //   console.log(this.openPaymentsService.client)
+        //   return
 
         default:
           return
