+++ src/background/services/index.ts
@@ -1,2 +1,3 @@
 export * from './BrowserEventsService'
 export * from './EventsService'
+export * from './OpenPaymentsService'
